+++ reckoner_dir/HistogramAnalyzer.cpp
@@ -4,7 +4,7 @@
 * This software is distributed under GNU GPL 3 license.
 *
 * Authors: Yun Heo, Maciej Dlugosz
-* Version: 2.0
+* Version: 2.1
 *
 */
 
@@ -25,7 +25,7 @@ bool HistogramAnalyzer::buildHistogram(const std::string& kmcDatabase, const uns
     }
 
     CKmerAPI kmer(kmcFileInfo.kmer_length);
-    float count;
+    uint32 count;
     unsigned iCount;
 
     histogram.clear();
